Refactor code for better embeddings
This commit bring a huge change in the code. The project has been focused on one single experiment, and the modules are going to change in order to favor that experiment.
This commit deeply modify the way the embeddings are computed: first, the embeddings are collected for each pair word/template, and after that they are aggregated with different parameter. This avoid to recompute everything from scratch for each experiment or testcase.